do not rerun precompute if we already have embeddings

+++ precompute_image_embeddings.py
@@ -5,30 +5,31 @@ import constants
 from PIL import Image
 from transformers import CLIPModel, CLIPProcessor
 
-loading_start = time.monotonic()
+if not os.path.isfile(constants.IMAGE_EMBEDDING_STORE_PATH):
+    loading_start = time.monotonic()
 
-model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32", cache_dir = constants.PRETRAINED_MODEL_CACHE_PATH)
-processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32", use_fast=True, cache_dir = constants.PRETRAINED_MODEL_CACHE_PATH)
+    model = CLIPModel.from_pretrained("openai/clip-vit-base-patch32", cache_dir = constants.PRETRAINED_MODEL_CACHE_PATH)
+    processor = CLIPProcessor.from_pretrained("openai/clip-vit-base-patch32", use_fast=True, cache_dir = constants.PRETRAINED_MODEL_CACHE_PATH)
 
-loading_end = time.monotonic()
-print(f"Loading Model Time: {(loading_end - loading_start): .2f} seconds")
+    loading_end = time.monotonic()
+    print(f"Loading Model Time: {(loading_end - loading_start): .2f} seconds")
 
 # Load images
-image_urls = []
-for root, dirs, files in os.walk("images"):
-    for name in files:
-        image_urls.append(os.path.join(root, name))
+    image_urls = []
+    for root, dirs, files in os.walk("images"):
+        for name in files:
+            image_urls.append(os.path.join(root, name))
 
-images = [
-    Image.open(image_url) for image_url in image_urls
-]
+    images = [
+        Image.open(image_url) for image_url in image_urls
+    ]
 
-imagePreprocessor = processor(images=images, return_tensors="pt")
-with torch.no_grad():
-    image_embeddings = model.get_image_features(**imagePreprocessor)
+    imagePreprocessor = processor(images=images, return_tensors="pt")
+    with torch.no_grad():
+        image_embeddings = model.get_image_features(**imagePreprocessor)
 
-image_embeddings = image_embeddings / image_embeddings.norm(
-    dim=-1, keepdim=True
-)
+    image_embeddings = image_embeddings / image_embeddings.norm(
+        dim=-1, keepdim=True
+    )
 
-torch.save({constants.IMAGE_EMBEDDINGS_KEY: image_embeddings, constants.IMAGE_URLS_KEY: image_urls}, constants.IMAGE_EMBEDDING_STORE_PATH)
+    torch.save({constants.IMAGE_EMBEDDINGS_KEY: image_embeddings, constants.IMAGE_URLS_KEY: image_urls}, constants.IMAGE_EMBEDDING_STORE_PATH)
